+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 - Bump CUDA base image version to 12.3.2
+- Add `cdi-cri` device list strategy. This uses the CDIDevices CRI field to request CDI devices instead of annotations.
 
 ### Version v0.15.0-rc.1
 - Import GPU Feature Discovery into the GPU Device Plugin repo. This means that